+++ docs/examples/docs-demo-six-sidebar-item-1.vue
@@ -1,0 +1,22 @@
+<template>
+<div>
+
+        <div style="width: 10em" id="six-sidebar-items-icon">
+          <six-sidebar-item icon="description">Data</six-sidebar-item>
+          <six-sidebar-item icon="history">History</six-sidebar-item>
+          <six-sidebar-item icon="account_balance">Transactions</six-sidebar-item>
+          <six-sidebar-item icon="upload">Upload</six-sidebar-item>
+          <six-sidebar-item icon="link" href="http://www.google.ch">Link</six-sidebar-item>
+        </div>
+      
+</div>
+</template>
+<style>
+
+</style>
+<script>
+export default {
+  name: 'docs-demo-six-sidebar-item-1',
+  mounted() {  }
+}
+</script>